+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.5.0] - January 10, 2020 (Kent, CT)
+### Changed
+- Changed all ``lite`` calls to ``sqlite3``. 
 
 ## [v0.4.4] - January 10, 2020 (Kent, CT)
 ### Added